Changed most doubles to decimal finally, got some sort of output from the multiItem.pack()method

+++ EventHorizon.Blazor.TypeScript.Interop.Generator-main/EventHorizon.Blazor.BabylonJS/Data/rectPos.cs
@@ -2,14 +2,14 @@ namespace EventHorizon.Blazor.BabylonJS.Data
 {
     public class rectPos
     {
-        public double[] startPos { get; set; }
+        public decimal[] startPos { get; set; }
 
-        public double[] endPos { get; set; }
+        public decimal[] endPos { get; set; }
 
         public rectPos()
         {
-            startPos = new double[3] { 0, 0, 0 };
-            endPos = new double[3] { 0, 0, 0 };
+            startPos = new decimal[3] { 0, 0, 0 };
+            endPos = new decimal[3] { 0, 0, 0 };
 
         }
     }
